Introduce ESLint and enforce it
Use the Maven frontend plugin to download NodeJS/npm and run eslint.

Some of the configuration bits are aligned with Jenkins core (such as
ECMA version 2022 or the list of available globals). The stylistic rules
are customized to align with the exisiting code base.

Lot of rules are ignored to prevent any code change.