Modernize the UI design
Replace the collapsible sections with a `<details>` HTML Element, this
has been available in all browsers since early 2020.  It defaults to
being collapsed and showing the `<summary>` content.  Toggling is
handled by the browser which let us get rid of the `doToggle()` script
(which also had duplicate logic to support IE 8).

The styling is inspired from Jenkins values and some shadow is casted at
the bottom of the sections to help differentiate them.