Ensure panel to be the last in the sidebar
Other plugins add sidebar elements and the order in which they are added
depends on how fast their `fetch()` requests are fulfiled.  The
Timestamper outline is slightly slower to respond and is most often put
last as a result.

Add an observer for any child changes made to the sidebar, when the
collapsible sections panel is not the last, move it after the last
panel ensuring it is always last.

Add Timestamper as a test dependency to ease testing when using hpi:run.