Add missing <ul> on top section
The first section started directly as a list item and missed an <ul>.
When later changing the outline to use the Jenkins panels styling, that
causes the first element to not have the necessary marging/padding and
have it sticked to the left border.

Make generateOutlineSection() to start with an <ul>, which adds the
missing one from the first section. Doing so also let us remove the ul
creation when inserting childs, since they are generated by
generateOutlineSection() and thus get an <ul>.