Make the panel sticky and remove obsolete code
The panel holding the section was made to stick to the top when the
window was scrolled. Nowadays this can be achieved with the CSS
property `position:sticky`.

Remove the JavaScript code attached to `window.onscroll`.
Drop the now unused div.scrollAttached

The panel is made to stick at 44px from the top, which is the value used
by the Timestamper plugin.

Before Jenkins 2.463/2.479.1, the pane is transparent and overlapping
two sticky had the texts overlayed which. Made it use a solid
background, thus when scrolling, the second panel hides the first one.