Fix jumping from sidebar link to a collapsed section
Follows-up c0b2fe11d37228, which broke this.

Before that, the server output this:

```
<div class=section ..>
  <div class=collapseHeader> .. <p onClick="doToggle(this)">..</p></div>
  <div class="collapsed">
  …
  </div>
</div>
```

And, script.js would prepend the `<a name>` destination to each
`div.section` element, which worked because the part that was
toggled is the div.collapsed/div.expanded child, not the section
itself.

After that, the server output this:

```
<details class=collapsingSection>
  <summary class=collapseHeader>..</summary>
  ..
  <div>
  </div>
</details>
```

Where `<details>` is entirely hidden by the browser, except for
the `<summary>` child. Any content outside of it the `summary` child
is thus part of the content that is toggled. The collapsible element
is thus `<details>`, not the `<div>`.

And, script.js still prepended `<a name>` to the same re-used
`sectionElt` variable, now pointing at `details.collapsingSection`,
thus making it in some strange way, toggleable content *above* the
summary and its toggle button.

Fix this by removing the element and simply setting an ID on the
`<details>` element itself.